+++ narrow_down/similarity_store.py
@@ -202,7 +202,17 @@ class SimilarityStore:
     async def insert(
         self, document: str, *, document_id: int = None, exact_part: str = None, data: str = None
     ) -> int:
-        """Index a new document."""
+        """Index a new document.
+
+        Args:
+            document: A document (as string to index).
+            document_id: Optional ID to assign to the document.
+            exact_part: Optional exact string to match when searching for the document.
+            data: Optional additional payload to save together with the document.
+
+        Returns:
+            The ID under which the document was indexed.
+        """
         tokens = self._tokenize_callable(document)
         fingerprint = self._minhasher.minhash(tokens)
         stored_doc = StoredDocument(
@@ -235,7 +245,34 @@ class SimilarityStore:
         await self._lsh.remove_by_id(document_id, check_if_exists)
 
     async def query(self, document: str, *, exact_part=None) -> Collection[StoredDocument]:
-        """Query all similar documents."""
+        """Query all similar documents.
+
+        Args:
+            document: A document to search similar items for.
+            exact_part: Part that should be exactly matched.
+
+        Returns:
+            A List of :obj:`~narrow_down.data_types.StoredDocument` objects with all elements
+            which are estimated to be above the similarity threshold.
+        """
         tokens = self._tokenize_callable(document)
         fingerprint = self._minhasher.minhash(tokens)
         return await self._lsh.query(fingerprint=fingerprint, exact_part=exact_part)
+
+    async def query_top_n(
+        self, n: int, document: str, *, exact_part=None
+    ) -> Collection[StoredDocument]:
+        """Query the top n similar documents.
+
+        Args:
+            n: The number of similar documents to retrieve.
+            document: A document to search similar items for.
+            exact_part: Part that should be exactly matched.
+
+        Returns:
+            A List of :obj:`~narrow_down.data_types.StoredDocument` objects with the n
+            elements which are most likely above the similarity threshold.
+        """
+        tokens = self._tokenize_callable(document)
+        fingerprint = self._minhasher.minhash(tokens)
+        return await self._lsh.query_top_n(n=n, fingerprint=fingerprint, exact_part=exact_part)