+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- SimilarityStore.top_n_query() now allows to find a limited number of most similar documents.
+
 ### Changed
 - SimilarityStore objects can now be created by a factory coroutine `create()` instead of
   calling first `__init__()` and then `initialize()`. This makes the usage of the class more